#3: Relocated constants, enums and struct definitions from WinUsbDevice.h to constants.h
