+++ XBOFS.win/XBOFS.win.vcxproj.filters
@@ -30,6 +30,9 @@
     <ClInclude Include="include\XBOFS.win\stdafx.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="include\XBOFS.win\constants.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <Filter Include="Resource Files">
